+++ dto/device_verification_response.go
@@ -1,5 +1,5 @@
 //
-// Copyright (C) 2019 Authlete, Inc.
+// Copyright (C) 2019-2020 Authlete, Inc.
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
@@ -45,4 +45,7 @@ type DeviceVerificationResponse struct {
 
 	//
 	ExpiresAt uint64 `json:"expiresAt,omitempty"`
+
+	//
+	Resources []string `json:"resources,omitempty"`
 }