+++ CHANGES.md
@@ -41,6 +41,9 @@ CHANGES
 - `DeviceCompleteRequest` struct
     * Added `IdtHeaderParams` field.
 
+- `DeviceVerificationResponse` struct
+    * Added `Resources` field.
+
 v1.0.5 (2019-09-25)
 -------------------